Fixes #4571: Move r31536 to v1.8 branch (fix abnormal code connection path in tcp oob)
---svn-pre-commit-ignore-below---

r31536 [[BR]]
Derived from patch provided by Artem, cleanup the "abnormal" code path for selecting TCP OOB modules to connect to a remote process. If we can't find a direct interface-to-address match, then assign all the provided addresses to the first available TCP module and let the normal failure process determine if the remote proc is truly reachable.

cmr=v1.8.2:reviewer=artpol:subject=fix abnormal code connection path in tcp oob


r31537 [[BR]]
Ensure we don't silently fail when unable to make a connection - bark pleasantly first.

Refs #4571


r31589 [[BR]]
Try again to get an error message printed when a daemon fails to successfully report back to mpirun. In this case, there is no guaranteed way for the daemon to output the error report itself - we don't have a connection back to the HNP, and we have tied stderr off to /dev/null (for good reasons). So the HNP has to detect the failure itself and report it.

The HNP can't know the precise reason, of course - all it knows is that the daemon failed. So output a generic error message that provides guidance on probable causes.

Refs #4571