Fixes #4608: Move r31688 to v1.8 branch (usnic: reduce sizeof(ompi_btl_usnic_addr_t) to 56)
---svn-pre-commit-ignore-below---

r31688 [[BR]]
usnic: reduce sizeof(ompi_btl_usnic_addr_t) to 56 bytes
    
Trivial struct re-ordering to eliminate holes in the middle of the
struct (although there's still a hole at the end) and reduce the
overall size of the struct from 64 to 56 bytes.  Also change mtu from
int to uint16_t; there was no need for it to be that large.

Reviewed by Dave Goodell

cmr=v1.8.2:reviewer=ompi-rm1.8