Fixes #4627: Move r31759 to v1.8 branch (Close the incoming socket if)
---svn-pre-commit-ignore-below---

r31759 [[BR]]
Close the incoming socket if we aren't going to accept it

cmr=v1.8.2:reviewer=rhc


r31763 [[BR]]
Just pick a module to handle the incoming connection if no direct interface is identified. Siegmar hit it because his IP/netmask is disjoint, but a router was able to make the connection.

Refs #4627