Fixes #4638: Move r31777 to v1.8 branch (opal: fix leaks)
NOTE: the part of this changeset that was supposed to apply to opal/mca/dstore/hash
could not be applied as that component does not exist in 1.8. Instead, it remains
in the opal/mca/db framework. A custom patch to move it across is required.


---svn-pre-commit-ignore-below---

r31777 [[BR]]
opal: fix leaks

Two leaks are fixed by this commit:

 - opal_dss.lookup_data_type returns an allocated string. Free it.

 - opal_ifaddrtokindex was leaking a struct addrinfo. Ensure that is
   released before returning.

cmr=v1.8.2:reviewer=rhc