Fixes #4530: Move r31419 to v1.8 branch (bcol/basesmuma: fix broken allgather algorithm)
---svn-pre-commit-ignore-below---

r31419 [[BR]]
bcol/basesmuma: fix broken allgather algorithm

The algorithm was failing ibm/collective/allgather and iallgather. I
cleaned up the code to eliminate duplicate code paths and tracked the
issue down to an error in the way extra nodes in the knomial exchange
are handled. The new code is more compact and has been tested with up
to 64 ranks with the ibm test suite.

cmr=v1.8.1:reviewer=manjugv