Fixes #4675: Move r31882 to v1.8 branch (plm/alps: better workaround for the)
---svn-pre-commit-ignore-below---

r31882 [[BR]]
plm/alps: better workaround for the noisy cray pmi implementation

This commit is a slightly better workaround to prevent mesages of
the form:
[unset]:_pmi_alps_get_apid:alps_app_lli_put_request failed
[unset]:_pmi_alps_get_appLayout:pmi_alps_get_apid returned with error: Bad file descriptor

It works by completely disabling PMI in the application process when using
mpirun. This should not be an issue for any apps.

cmr=v1.8.2:reviewer=rhc