Fixes #4697: Move r31955 to v1.8 branch (Fix the Fortran issue identified)
---svn-pre-commit-ignore-below---

r31955 [[BR]]
Fix the Fortran issue identified by Akan Sang Loon. The dist graph
is really special as the weights can be one of the following three
values (NULL, EMPTY or some legal value). As such, we need a complex
if to correctly convert the Fortran value to the corresponding C
value. Thus, always defining the c_ array is the simplest and most
straighforward approach.

cmr=v1.8.2:reviewer=jsquyres