Fixes #4730: Move r32015 to v1.8 branch (osc/rdma: fix accumulate fragment size)
---svn-pre-commit-ignore-below---

r32015 [[BR]]
osc/rdma: fix accumulate fragment size calculation

The wrong type was used when calculating the amount of space needed
for an accumulate fragment. Fixed the calculation and took the
opportunity to eliminate the get_acc header as it is identical to the
acc header.

This fixes #4719 and #4718

Tracking these fixes for 1.8.2 in this CMR.

Throwing this to Brad for review as he is the one who ran into the issue.

cmr=v1.8.2:reviewer=bbenton

r32016 [[BR]]
osc/rdma: fix hang when accumulating with MPI_REPLACE

The replace callback did not increment the incoming frag counter. This
leads to a hang during synchronization. This commit adds the increment
and also puts the request on the garbage collection list to fix a leak.

This fixes a hang found when running the mpich test suite.

cmr=v1.8.2:reviewer=bbenton